Storage improvements

+++ src/Storage/BotDB.php
@@ -11,7 +11,7 @@
 namespace Bot\Storage;
 
 use Bot\Exception\BotException;
-use Bot\Helper\DebugLog;
+use Bot\Helper\Debug;
 use Longman\TelegramBot\Exception\TelegramException;
 use Longman\TelegramBot\DB;
 use PDO;
@@ -27,61 +27,32 @@ use PDOException;
 class BotDB extends DB
 {
     /**
-     * This is 'proxy' function to server actions
+     * SQL to create database structure
      *
-     * @param $action
-     * @param $id
-     * @param array  $data
-     *
-     * @return array|bool|mixed
-     * @throws BotException
+     * @var string
      */
-    public static function action($action, $id, $data = [])
-    {
-        if (empty($action)) {
-            throw new BotException('Action is empty!');
-        }
+    private static $structure = 'CREATE TABLE IF NOT EXISTS `storage` (
+        `id` CHAR(100) COMMENT "Unique identifier for this entry",
+        `data` TEXT NOT NULL) COMMENT "Stored data",
+        `created_at` timestamp NULL DEFAULT NULL) COMMENT "Entry creation date",
+        `updated_at` timestamp NULL DEFAULT NULL) COMMENT "Entry update date",
 
-        if (empty($id)) {
-            throw new BotException('Id is empty!');
-        }
-
-        self::initializeStorage();
-
-        switch ($action) {
-            default:
-            case 'read':
-                return self::selectFromStorage($id);
-            case 'save':
-                if (empty($data)) {
-                    throw new BotException('Data is empty!');
-                }
-
-                return self::insertToStorage($id, $data);
-            case 'lock':
-                return self::lockStorage($id);
-            case 'unlock':
-                return self::unlockStorage($id);
-            case 'list':
-                return self::listFromStorage($id);
-            case 'remove':
-                return self::deleteFromStorage($id);
-        }
-    }
+        PRIMARY KEY (`id`)
+    );';
 
     /**
-     * Initialize 'storage' table
+     * Initialize PDO connection and 'storage' table
      */
-    private static function initializeStorage()
+    public static function initializeStorage()
     {
         if (!defined('TB_STORAGE')) {
             define('TB_STORAGE', self::$table_prefix . 'storage');
 
-            if (!file_exists($structure_check = VAR_PATH . '/db_structure_created') && file_exists($structure = ROOT_PATH . '/structure.sql')) {
-                DebugLog::log('Creating database structure...');
+            if (!file_exists($structure_check = VAR_PATH . '/db_structure_created')) {
+                Debug::log('Creating database structure...');
 
                 try {
-                    if ($result = self::$pdo->query(file_get_contents($structure))) {
+                    if ($result = self::$pdo->query(self::$structure)) {
                         if (!is_dir(VAR_PATH)) {
                             mkdir(VAR_PATH, 0755, true);
                         }
@@ -95,6 +66,8 @@ class BotDB extends DB
                 }
             }
         }
+
+        return true;
     }
 
     /**
@@ -105,10 +78,14 @@ class BotDB extends DB
      * @return array|bool|mixed
      * @throws BotException
      */
-    private static function selectFromStorage($id)
+    public static function selectFromStorage($id)
     {
         if (!self::isDbConnected()) {
             return false;
+        }
+
+        if (empty($id)) {
+            throw new BotException('Id is empty!');
         }
 
         try {
@@ -139,10 +116,18 @@ class BotDB extends DB
      * @return bool
      * @throws BotException
      */
-    private static function insertToStorage($id, $data)
+    public static function insertToStorage($id, $data)
     {
         if (!self::isDbConnected()) {
             return false;
+        }
+
+        if (empty($id)) {
+            throw new BotException('Id is empty!');
+        }
+
+        if (empty($data)) {
+            throw new BotException('Data is empty!');
         }
 
         try {
@@ -170,74 +155,6 @@ class BotDB extends DB
     }
 
     /**
-     * Lock the row to prevent another process modifying it
-     *
-     * @TODO this is really bad way of doing it currently, forces running this project just on one dyno... (they do not share filesystem)
-     * @TODO maybe return `self::$pdo->beginTransaction();` could be somehow used here?
-     *
-     * @param $id
-     *
-     * @return bool
-     */
-    private static function lockStorage($id)
-    {
-        if (!is_dir(VAR_PATH . '/tmp')) {
-            mkdir(VAR_PATH . '/tmp', 0755, true);
-        }
-
-        return flock(fopen(VAR_PATH . '/tmp/' . $id .  '.json', "a+"), LOCK_EX);
-    }
-
-    /**
-     * Unlock the row after
-     *
-     * @TODO this is really bad way of doing it currently, forces running this project just on one dyno... (they do not share filesystem)
-     * @TODO maybe `return self::$pdo->commit();` could be somehow used here?
-     *
-     * @param $id
-     *
-     * @return bool
-     */
-    private static function unlockStorage($id)
-    {
-        return flock(fopen(VAR_PATH . '/tmp/' . $id .  '.json', "a+"), LOCK_UN) && unlink(VAR_PATH . '/tmp/' . $id .  '.json');
-    }
-
-    /**
-     * Select multiple data from the database
-     *
-     * @param int $time
-     *
-     * @return array|bool
-     * @throws BotException
-     */
-    private static function listFromStorage($time = 0)
-    {
-        if ($time < 0) {
-            throw new BotException('Time cannot be a negative number!');
-        }
-
-        try {
-            $sth = self::$pdo->prepare('
-                SELECT * FROM `' . TB_STORAGE . '`
-                WHERE `updated_at` < :date
-            ');
-
-            $date = self::getTimestamp(strtotime('-' . $time . ' seconds'));
-
-            $sth->bindParam(':date', $date, PDO::PARAM_STR);
-
-            if ($result = $sth->execute()) {
-                return $sth->fetchAll(PDO::FETCH_ASSOC);
-            } else {
-                return $result;
-            }
-        } catch (PDOException $e) {
-            throw new BotException($e->getMessage());
-        }
-    }
-
-    /**
      * Delete data from storage
      *
      * @param $id
@@ -251,6 +168,10 @@ class BotDB extends DB
             return false;
         }
 
+        if (empty($id)) {
+            throw new BotException('Id is empty!');
+        }
+
         try {
             $sth = self::$pdo->prepare('
                 DELETE FROM `' . TB_STORAGE . '`
@@ -260,6 +181,82 @@ class BotDB extends DB
             $sth->bindParam(':id', $id, PDO::PARAM_STR);
 
             return $sth->execute();
+        } catch (PDOException $e) {
+            throw new BotException($e->getMessage());
+        }
+    }
+
+    /**
+     * Lock the row to prevent another process modifying it
+     *
+     * @TODO this is really bad way of doing it currently, forces running this project just on one dyno... (they do not share filesystem)
+     * @TODO maybe `self::$pdo->beginTransaction();` could be somehow used here?
+     *
+     * @param $id
+     *
+     * @return bool
+     */
+    public static function lockStorage($id)
+    {
+        if (empty($id)) {
+            throw new BotException('Id is empty!');
+        }
+
+        if (!is_dir(VAR_PATH . '/tmp')) {
+            mkdir(VAR_PATH . '/tmp', 0755, true);
+        }
+
+        return flock(fopen(VAR_PATH . '/tmp/' . $id .  '.json', "a+"), LOCK_EX);
+    }
+
+    /**
+     * Unlock the row after
+     *
+     * @TODO this is really bad way of doing it currently, forces running this project just on one dyno... (they do not share filesystem)
+     * @TODO maybe `self::$pdo->commit();` could be somehow used here?
+     *
+     * @param $id
+     *
+     * @return bool
+     */
+    public static function unlockStorage($id)
+    {
+        if (empty($id)) {
+            throw new BotException('Id is empty!');
+        }
+
+        return flock(fopen(VAR_PATH . '/tmp/' . $id .  '.json', "a+"), LOCK_UN) && unlink(VAR_PATH . '/tmp/' . $id .  '.json');
+    }
+
+    /**
+     * Select multiple data from the database
+     *
+     * @param int $time
+     *
+     * @return array|bool
+     * @throws BotException
+     */
+    public static function listFromStorage($time = 0)
+    {
+        if ($time < 0) {
+            throw new BotException('Time cannot be a negative number!');
+        }
+
+        try {
+            $sth = self::$pdo->prepare('
+                SELECT * FROM `' . TB_STORAGE . '`
+                WHERE `updated_at` <= :date
+            ');
+
+            $date = self::getTimestamp(strtotime('-' . $time . ' seconds'));
+
+            $sth->bindParam(':date', $date, PDO::PARAM_STR);
+
+            if ($result = $sth->execute()) {
+                return $sth->fetchAll(PDO::FETCH_ASSOC);
+            } else {
+                return $result;
+            }
         } catch (PDOException $e) {
             throw new BotException($e->getMessage());
         }
